+++ src/data/arbitrary/types/mod.rs
@@ -2,5 +2,6 @@ mod amino_acids;
 mod cigar;
 mod nucleotides;
 
+pub use amino_acids::*;
 pub use cigar::*;
 pub use nucleotides::*;