+++ CHANGELOG.md
@@ -18,6 +18,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `SequenceReadable` getter trait to record types and sequence types
 - Added `nw_score_from_path` for rescoring a global alignment (behind `alignment-diagnostics` feature gate)
 - Introduced the `AlignmentAndSeqs` arbitrary wrapper (behind `fuzzing` feature gate)
+- Added `AminoAcidsIupacX` arbitrary wrapper (behind `fuzzing` feature gate)
 - Added `visit_params` methods to pHMMs as a diagnostic tool for inspecting the parameters that are visited along a given path (behind `dev-phmm` and `alignment-diagnostics` feature gates)
 - Added the ability to remove layers from a pHMM (behind `dev-phmm` and `alignment-diagnostics` feature gates)
 - Added random downsampling for sized and non-sized iterators using `MethodDSampler`, `method_l`, and `BernoulliSampler`